improves error checking in queries

+++ controllers/queries.js
@@ -14,7 +14,7 @@ module.exports = {
     var setup = req.body.setup;
     var interests = req.body.interests;
     var endTime = convertToMS( availableTime );
-    console.log( 'Received: ', username, availableTime, setup, interests );
+    if ( !username || !availableTime ) return res.status( 400 ).send( { status: 400, error: "Missing parameters." });
     // create a new post
     var newPost = Post({
       postTime: new Date().getTime(),
@@ -45,6 +45,7 @@ module.exports = {
 
   deletePost: ( req, res ) => {
     var id = new ObjectId( req.params.id );
+    if ( !ObjectID.isValid(id) ) return res.status( 400 ).send( {error: "Invalid id." })
     Post
       .findOneAndRemove( { '_id': id })
       .exec( () => res.redirect( 'http://pairboard.surge.sh' ));
@@ -69,6 +70,7 @@ module.exports = {
   // DELETE /posts/:id
   deleteOnePost: (req, res) => {
     var id = new ObjectId(req.params.id);
+    if ( !ObjectID.isValid(id) ) return res.status( 400 ).send( {error: "Invalid id." })
     Post
       .findOneAndRemove( { '_id': id })
       .then( ( data ) => {
